conversion from chrono

+++ src/constants.rs
@@ -4,8 +4,8 @@ pub(crate) const SECONDS_IN_ONE_SECOND: u32 =        1;
 pub(crate) const SECONDS_IN_ONE_MINUTE: u32 =       60;
 pub(crate) const SECONDS_IN_ONE_HOUR  : u32 =     3600;
 pub(crate) const SECONDS_IN_ONE_DAY   : u32 =    86400;
-pub(crate) const SECONDS_IN_ONE_WEEK : u32 =   604800;
-pub(crate) const SECONDS_IN_ONE_YEAR  : u32 = 31557600;
+pub(crate) const SECONDS_IN_ONE_WEEK  : u32 =   604800;
+pub(crate) const SECONDS_IN_ONE_YEAR  : u32 = 31557600; // 365.25 days!
 
 pub(crate) const SECONDS_IN_ONE: [u32; 6] = [
     SECONDS_IN_ONE_SECOND,
@@ -34,7 +34,7 @@ pub(crate) const SM: [u32; 14] = [ 0,
 ];
 
 /// days in months
-pub(crate) const MDAYS: [u32; 13] = [0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
+pub(crate) const MDAYS: [u8; 13] = [0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
 /// seconds in a day: 24*60*60
 pub(crate) const DAYLEN: u32 = 86400;
 /// seconds between 2000-01-01 and 1970-01-01
